sedac/v9/production/requirements.txt
# SEDAC V9.0 Production Requirements

# Core
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Quantization
bitsandbytes>=0.41.0
auto-gptq>=0.5.0

# Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Monitoring
prometheus-client>=0.19.0

# Data
datasets>=2.15.0
tokenizers>=0.15.0

# Utils
tqdm>=4.66.0
pyyaml>=6.0.0
numpy>=1.24.0

# Optional: Flash Attention
# flash-attn>=2.3.0

# Optional: vLLM integration
# vllm>=0.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
